% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actives_datastat.R
\name{actives_datastat}
\alias{actives_datastat}
\title{-------- Fonction actives_datastat() --------}
\usage{
actives_datastat()
}
\description{
Cette fonction retourne, jour par jour, le nombre de \strong{nouvelles pharmacies adhérentes}
(actives) au service \emph{DataStat}, ainsi que la somme cumulée depuis 2002‑01‑01.
}
\details{
##Usage

actives_datastat()
\subsection{Résultat}{

Un \code{data.frame} (tibble) avec les colonnes:
\itemize{
\item \code{n_auto_adhpha}          : identifiant de la pharmacie (NA si aucune arrivée ce jour‑là)
\item \code{date_complete}          : date d’adhésion (AAAA/MM/JJ)
\item \code{Nb_Actif_journalier}    : nombre de nouveaux adhérents ce jour‑là
\item \code{somme_cumulative_Actif} : cumul des adhérents depuis le 1ᵉʳjanvier 2002
}
}

\subsection{Exemple}{

actives_datastat()
}
}
