% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_connect_db.R
\name{auto_connect_db}
\alias{auto_connect_db}
\title{Connexion automatique à une base de données SQL}
\usage{
auto_connect_db(database = "OSP_DATASTAT")
}
\arguments{
\item{database}{Nom de la base de données (par défaut: "OSP_DATASTAT")}
}
\value{
Un objet de connexion DBI
}
\description{
Cette fonction permet de se connecter automatiquement à une base de données SQL
en utilisant les variables d'environnement configurées.
}
\examples{
\dontrun{
con <- auto_connect_db()
DBI::dbListTables(con)
DBI::dbDisconnect(con)
}
}
