% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_interactive_departements.R
\name{carte_interactive_departements}
\alias{carte_interactive_departements}
\title{Carte interactive des départements de France avec \code{leaflet}}
\usage{
carte_interactive_departements(
  base,
  var_code_dept,
  var_valeurs,
  titre_carte = "Carte Interactive des Départements de France",
  couleur_palette = "YlOrRd",
  logo_path = NULL,
  signe_unite = 0,
  valeurs_fixes_supp = NULL
)
}
\arguments{
\item{base}{Un \code{data.frame} contenant les données des départements avec les valeurs associées.}

\item{var_code_dept}{Nom de la colonne contenant les codes des départements.}

\item{var_valeurs}{Vecteur de noms de colonnes des variables à afficher en info-bulle (\code{tooltip}).
La \strong{première variable} est utilisée pour la coloration de la carte.}

\item{titre_carte}{Titre affiché sur la carte (\code{default = "Carte Interactive des Départements de France"}).}

\item{couleur_palette}{Palette de couleurs pour \code{leaflet} (\code{default = "YlOrRd"}).}

\item{logo_path}{Chemin ou URL du logo à afficher en haut à droite (\code{default = NULL}).}

\item{signe_unite}{Indique l'unité à afficher pour la variable principale (\code{default = 0}).
- \code{0} : Pas d'unité
- \code{1} : Affiche \verb{€}
- \code{2} : Affiche \verb{\%}}

\item{valeurs_fixes_supp}{Vecteur de noms de colonnes des variables supplémentaires à afficher avec la valeur principale (\code{default = NULL}).}
}
\value{
Un objet \code{leaflet} affichant la carte interactive des départements de France.
}
\description{
Cette fonction génère une carte interactive des départements de France en utilisant les codes des départements comme clé d'association.
Les couleurs sont basées sur une variable principale et peuvent être personnalisées.
Il est aussi possible d'afficher des valeurs supplémentaires en info-bulle.
}
