% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ean.R
\name{ean}
\alias{ean}
\title{-------- Fonction ean() --------}
\usage{
ean()
}
\description{
Cette fonction récupère les codes EAN associés aux articles.
}
\details{
##Usage

ean()
\subsection{Résultat}{

Un data.frame (tibble) avec les colonnes suivantes :
\itemize{
\item \code{n_auto_artic} : identifiant de l'article
\item \code{code_ean}     : code EAN correspondant
}
}

\subsection{Exemple}{

ean()
}
}
