% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fournisseurs.R
\name{fournisseurs}
\alias{fournisseurs}
\title{-------- Fonction fournisseurs() --------}
\usage{
fournisseurs()
}
\description{
Cette fonction récupère la liste des fournisseurs référencés dans la base.
}
\details{
##Usage

fournisseurs()
\subsection{Résultat}{

Un data.frame (tibble) avec les colonnes suivantes :
\itemize{
\item \code{n_auto_adhfour} : identifiant unique du fournisseur
\item \code{nom_adhfour}    : nom du fournisseur
\item \code{cp_adhfour}     : code postal
\item \code{ville_adhfour}  : ville
}
}

\subsection{Exemple}{

fournisseurs()
}
}
