% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentation.R
\name{frequentation_journaliere}
\alias{frequentation_journaliere}
\title{Calcul de la fréquentation journalière moyenne}
\usage{
frequentation_journaliere(n_auto, date_debut, date_fin)
}
\arguments{
\item{n_auto}{Identifiant de la pharmacie (\code{n_auto_adhpha_ecollect})}

\item{date_debut}{Période de début au format \code{YYYYMM} (ex : 202401)}

\item{date_fin}{Période de fin au format \code{YYYYMM} (ex : 202404)}
}
\value{
Un data.frame avec :
\itemize{
\item \code{n_auto} : identifiant de la pharmacie
\item \code{frequentation_moyenne_jour} : fréquentation moyenne journalière
}
}
\description{
Cette fonction calcule la \strong{fréquentation moyenne journalière} d'une pharmacie
(nombre moyen de clients par jour), sur une période personnalisée.
}
\examples{
\dontrun{
frequentation_journaliere(123456, 202401, 202404)
}

}
