% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_month.R
\name{get_previous_month}
\alias{get_previous_month}
\title{get_previous_month - Obtenir un mois précédent sous format YYYYmm}
\usage{
get_previous_month(n = 0)
}
\arguments{
\item{n}{Un entier représentant le nombre de mois à remonter (par défaut 0 pour le mois précédent).}
}
\value{
Une chaîne de caractères représentant l'année et le mois au format YYYYmm.
}
\description{
Cette fonction permet d'obtenir la date du mois précédent par défaut,
ou un nombre de mois avant spécifié par l'utilisateur.
}
\examples{
get_previous_month()   # Mois précédent
get_previous_month(1)  # Deux mois avant
get_previous_month(12) # Douze mois avant

}
