% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventes_suspectes.R
\name{get_ventes_suspectes}
\alias{get_ventes_suspectes}
\title{Identifier les pharmacies avec des ventes suspectes}
\usage{
get_ventes_suspectes(periode, seuil_suspect = 999)
}
\arguments{
\item{periode}{Un entier représentant la période au format YYYYMM (ex: 202505)}

\item{seuil_suspect}{Seuil de quantité considérée comme suspecte (défaut: 999)}
}
\value{
Un data.frame contenant les colonnes:
\itemize{
\item dateheure_ecollect
\item n_auto_adhpha
\item LGO
\item typ_vte_ecollect
\item n_auto_artic
\item nom_artic
\item quantite_vendue
}
}
\description{
Cette fonction identifie les pharmacies ayant des quantités vendues
suspectes pour une période donnée.
}
