% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupement.R
\name{groupement}
\alias{groupement}
\title{-------- Fonction groupement() --------}
\usage{
groupement()
}
\description{
Cette fonction retourne la liste des groupements de pharmacies
ainsi que les identifiants des pharmacies associées.
}
\details{
##Usage

groupement()
\subsection{Résultat}{

Un data.frame (tibble) avec les colonnes suivantes :
\itemize{
\item \code{n_auto_adhpha} : identifiant de la pharmacie
\item \code{n_auto_adhgrp} : identifiant du groupement
\item \code{nom_adhgrp}    : nom du groupement
}
}

\subsection{Exemple}{

groupement()
}
}
