% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactifs_datastat.R
\name{inactifs_datastat}
\alias{inactifs_datastat}
\title{-------- Fonction inactifs_datastat() --------}
\usage{
inactifs_datastat()
}
\description{
Cette fonction retourne, jour par jour, le nombre de \strong{pharmacies désactivées (inactives)}
du service \emph{DataStat}, ainsi que leur cumul depuis le 01/01/2002.
}
\details{
##Usage

inactifs_datastat()
\subsection{Résultat}{

Un \code{data.frame} (tibble) avec les colonnes suivantes :
\itemize{
\item \code{n_auto_adhpha}            : identifiant de la pharmacie (NA si aucune sortie ce jour-là)
\item \code{date_complete}            : date de sortie (AAAA/MM/JJ)
\item \code{Nb_Inactif_journalier}    : nombre de pharmacies devenues inactives ce jour-là
\item \code{somme_cumulative_Inactif} : cumul total d’inactivations depuis 2002
}
}

\subsection{Exemple}{

inactifs_datastat()
}
}
