% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infos_departement.R
\name{infos_departement}
\alias{infos_departement}
\title{Infos sur les départements français (via sf)}
\usage{
infos_departement()
}
\value{
Un data.frame avec le nom, le code, la longitude et la latitude du centroïde
}
\description{
Cette fonction télécharge et lit le fichier GeoJSON des départements français
via le package sf.
}
\examples{
\dontrun{
  df <- infos_departement()
  head(df)
}
}
