% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle_article.R
\name{libelle_article}
\alias{libelle_article}
\title{-------- Fonction libelle_article() --------}
\usage{
libelle_article()
}
\description{
Cette fonction retourne les informations principales d'un article :
identifiant, famille, marque, type, taux de TVA, nom, et fournisseur.
}
\details{
##Usage

libelle_article()
\subsection{Résultat}{

Un data.frame (tibble) avec les colonnes suivantes :
\itemize{
\item \code{n_auto_artic}   : identifiant article
\item \code{n_auto_famille} : identifiant de la famille
\item \code{lib_famille}    : nom de la famille
\item \code{marque}         : marque du produit
\item \code{type_artic}     : type d'article
\item \code{tva_artic}      : taux de TVA lié au produit
\item \code{nom_artic}      : nom de l'article
\item \code{adhfour}        : code fournisseur
}
}

\subsection{Exemple}{

libelle_article()
}
}
