% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle_tva.R
\name{libelle_tva}
\alias{libelle_tva}
\title{Libellés des taux de TVA}
\usage{
libelle_tva(syndic_only = TRUE)
}
\arguments{
\item{syndic_only}{Si \code{TRUE} (par défaut), on ne garde que les lignes actives.}
}
\value{
Un tableau (tibble) avec les colonnes : \code{code_tva}, \code{lib_tva}, \code{taux_tva}.
}
\description{
Cette fonction retourne la liste des taux de TVA disponibles dans la base.
}
\examples{
libelle_tva()          # taux actifs uniquement
libelle_tva(FALSE)     # tous les taux

}
