% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periode_window.R
\name{periode_window}
\alias{periode_window}
\title{-------- Fonction periode_window() --------}
\usage{
periode_window(n = NULL, date_debut = NULL, date_fin = NULL)
}
\description{
Cette fonction retourne une table des périodes (AAAAMM) avec le nom du mois
correspondant, soit sur la base d'une fenêtre glissante en mois, soit entre deux dates données.
}
\details{
##Usage

periode_window(n = 6)
periode_window(date_debut = 202301, date_fin = 202312)
\subsection{Paramètres}{
\itemize{
\item \code{n} : entier. Nombre de mois à retourner en arrière à partir du mois précédent (par défaut : NULL)
\item \code{date_debut} : entier (AAAAMM), début de période si \code{n} est NULL.
\item \code{date_fin} : entier (AAAAMM), fin de période si \code{n} est NULL.
}
}

\subsection{Résultat}{

Un data.frame avec deux colonnes :
\itemize{
\item \code{Periode} : code période au format AAAAMM
\item \code{Mois}    : nom du mois en français
}
}

\subsection{Exemple}{

periode_window(5)
periode_window(date_debut = 202201, date_fin = 202212)
}
}
