% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacies_infos.R
\name{pharmacies_infos}
\alias{pharmacies_infos}
\title{Informations sur les pharmacies d’un groupement (ou toutes)}
\usage{
pharmacies_infos(groupement = NULL)
}
\arguments{
\item{groupement}{Entier (optionnel). Identifiant du groupement.
Par défaut (\code{NULL}), la fonction renvoie toutes les pharmacies.}
}
\value{
Un data.frame avec les colonnes :
\itemize{
\item \code{n_auto_adhpha} : identifiant de la pharmacie
\item \code{PHARMACIE} : nom de la pharmacie
\item \code{CIP} : code CIP
\item \code{nom_region} : région
\item \code{nom_ville} : ville
\item \code{cp_ville} : code postal
\item \code{mail} : email
}
}
\description{
Cette fonction récupère les informations principales des pharmacies,
avec ou sans filtrage par groupement.
}
\examples{
\dontrun{
pharmacies_infos()        # Toutes les pharmacies
pharmacies_infos(164)     # Pharmacies du groupement 164
}

}
