% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiel_achat.R
\name{potentiel_achat}
\alias{potentiel_achat}
\title{potentiel_achat}
\usage{
potentiel_achat(
  n_auto_adhgrp = NULL,
  date_debut,
  date_fin,
  n_auto_adhfour_vect
)
}
\arguments{
\item{n_auto_adhgrp}{(optionnel) Identifiant du groupement de pharmacies (entier). Si NULL, récupère toutes les pharmacies.}

\item{date_debut}{Date de début au format AAAAMM (entier).}

\item{date_fin}{Date de fin au format AAAAMM (entier).}

\item{n_auto_adhfour_vect}{Vecteur des identifiants des fournisseurs (vecteur d'entiers).}
}
\value{
Un data.frame contenant les colonnes Pharmacie, cip, ville, et pour chaque fournisseur sélectionné :
POTENTIEL \if{html}{\out{<NOM FOURNISSEUR>}} n°1 en PFHT et REMISE PONDEREE \if{html}{\out{<NOM FOURNISSEUR>}}
}
\description{
Calcule le potentiel d’achat et la remise pondérée par pharmacie, cip et ville pour des fournisseurs donnés.
}
