% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_completude.R
\name{stat_completude}
\alias{stat_completude}
\title{Statistiques de complétude mensuelle par adhérent}
\usage{
stat_completude()
}
\value{
Un data.frame détaillé avec la complétude mensuelle par adhérent.
}
\description{
Cette fonction exécute une requête SQL intégrée sur la base \code{os_stat_artic},
calcule le nombre de mois consécutifs de données disponibles par adhérent,
et retourne la table complète de complétude sans le pourcentage.
}
\examples{
df_stat <- stat_completude()
}
