% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tranche_ mdl.R
\name{stat_tranche_mdl}
\alias{stat_tranche_mdl}
\title{Statistiques par tranche MDL sur une période donnée}
\usage{
stat_tranche_mdl(date_debut, date_fin)
}
\arguments{
\item{date_debut}{Période de début au format numérique \code{YYYYMM} (ex: 202306)}

\item{date_fin}{Période de fin au format numérique \code{YYYYMM} (ex: 202406)}
}
\value{
Un data.frame avec les colonnes : \code{periode}, \code{Tranche_MDL}, \code{CA}, \code{volume}.
}
\description{
Cette fonction interroge la base \code{os_stat_artic} en joignant la table
\code{CEPS_Prix} pour calculer les statistiques de CA et de volume par
tranche MDL. Elle prend en paramètre une période de début et une période de fin
(au format \code{YYYYMM}), et retourne un tableau agrégé par mois et tranche MDL.
}
\examples{
df_mdl <- stat_tranche_mdl(202301, 202405)
head(df_mdl)
}
