% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tva.R
\name{stat_tva}
\alias{stat_tva}
\title{Statistiques de ventilation par taux de TVA}
\usage{
stat_tva(date_debut, date_fin)
}
\arguments{
\item{date_debut}{Un entier au format YYYYMM représentant la période de début (ex: 202406)}

\item{date_fin}{Un entier au format YYYYMM représentant la période de fin (ex: 202505)}
}
\value{
Un data.frame contenant les colonnes : periode, tva_artic, CA, volume
}
\description{
Cette fonction récupère les statistiques de chiffre d'affaires (CA) et volume
des articles vendus ventilés par taux de TVA, pour les pharmacies complètes
entre deux périodes spécifiées.
}
\examples{
\dontrun{
stat_tva(202406, 202505)
}

}
