% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistique_article.R
\name{statistique_article}
\alias{statistique_article}
\title{Obtenir les statistiques d'articles sur une période définie}
\usage{
statistique_article(articles, date_debut, date_fin)
}
\arguments{
\item{articles}{Vecteur d'identifiants article (entiers).}

\item{date_debut}{Entier. Date de début au format AAAAMM (ex : 202306).}

\item{date_fin}{Entier. Date de fin au format AAAAMM (ex : 202505).}
}
\value{
Un data.frame avec une ligne par combinaison article/pharmacie/période,
contenant les colonnes :
- \code{periode}
- \code{n_auto_adhpha_artic}
- \code{n_auto_artic_artic}
- \code{CA_HT}
- \code{px_achat}
- \code{qte_vendu}
}
\description{
La fonction \code{statistique_article()} permet de récupérer les chiffres de vente
(CA HT, prix d'achat cumulé, quantité vendue) d'un ou plusieurs articles,
sur une période définie, en excluant les pharmacies incomplètes.
}
\examples{
statistique_article(c(344734), 202306, 202505)

}
