% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistique_famille.R
\name{statistique_famille}
\alias{statistique_famille}
\title{-------- Fonction statistique_famille() --------}
\usage{
statistique_famille(date_debut, date_fin)
}
\description{
Cette fonction retourne les statistiques des ventes par famille de produits
(CA HT, quantité vendue) sur une période donnée.
}
\details{
##Usage

statistique_famille(date_debut, date_fin)
\subsection{Résultat}{

Un data.frame (tibble) avec les colonnes suivantes :
\itemize{
\item \code{periode}        : période de référence (AAAAMM)
\item \code{n_auto_famille} : identifiant de la famille
\item \code{CA_HT}          : chiffre d'affaires HT (arrondi à 2 décimales)
\item \code{qte_vendu}      : quantité vendue
}
}

\subsection{Exemple}{

statistique_famille(202306, 202505)
}
}
