% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistique_type_article.R
\name{statistique_type_article}
\alias{statistique_type_article}
\title{-------- Fonction statistique_type_article() --------}
\usage{
statistique_type_article(date_debut, date_fin)
}
\description{
Cette fonction retourne les statistiques des ventes par type d'article
(ex : Princeps, Générique, Biosimilaire...) et selon leur caractère
substituable ou non, sur une période donnée.
}
\details{
##Usage

statistique_type_article(date_debut, date_fin)
\subsection{Résultat}{

Un data.frame (tibble) avec les colonnes suivantes :
\itemize{
\item \code{periode}      : période au format AAAAMM
\item \code{type_artic}   : type de l'article (Princeps, Biosimilaire, etc.)
\item \code{Subtitution}  : "Substituable" ou "Non substituable"
\item \code{CA_HT}        : chiffre d'affaires HT (arrondi 2 décimales)
\item \code{qte_vendu}    : quantité totale vendue
}
}

\subsection{Exemple}{

statistique_type_article(202306, 202505)
}
}
