% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executer_sql.R
\name{executer_sql}
\alias{executer_sql}
\title{Exécuter une requête SQL}
\usage{
executer_sql(query, params = NULL, database = "OSP_DATASTAT")
}
\arguments{
\item{query}{Chaîne de caractères contenant la requête SQL}

\item{params}{Liste optionnelle de paramètres pour les requêtes paramétrées}

\item{database}{Nom de la base de données (par défaut: "OSP_DATASTAT")}
}
\value{
Résultat de la requête (data frame)
}
\description{
Cette fonction permet d'exécuter facilement une requête SQL sur la base de données
en établissant automatiquement la connexion.
}
\examples{
\dontrun{
resultats <- executer_sql("SELECT * FROM ma_table WHERE id < ?", params = list(100))
}
}
