% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_complete.R
\name{pharma_complete_12}
\alias{pharma_complete_12}
\title{Obtenir les pharmacies complètes sur une période définie}
\usage{
pharma_complete_12(date_debut, date_fin)
}
\arguments{
\item{date_debut}{Entier. Date de début au format \code{YYYYMM} (ex : 202401).}

\item{date_fin}{Entier. Date de fin au format \code{YYYYMM} (ex : 202412). Doit être ≥ à \code{date_debut}.}
}
\value{
Un \code{data.frame} listant les pharmacies complètes sur la période indiquée, avec leurs données consolidées.
}
\description{
La fonction \code{pharma_complete_12()} retourne les pharmacies considérées comme \strong{complètes}
sur une période donnée (définie par une date de début et une date de fin au format \code{YYYYMM}),
ainsi que leurs indicateurs économiques principaux (CA HT, CA TTC, nombre de ventes, quantités vendues).
}
\details{
\itemize{
\item Les noms des colonnes de sortie sont fixes (non dynamiques).
}
}
\examples{
\dontrun{
pharma_complete_12(202401, 202412)
}

}
