% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_complete.R
\name{pharma_complete_12}
\alias{pharma_complete_12}
\title{Obtenir les pharmacies complètes pour une année donnée}
\usage{
pharma_complete_12(n = 0)
}
\arguments{
\item{n}{Entier (défaut = 0). Nombre d'années à reculer depuis l'année actuelle - 1.}
}
\value{
Un data.frame contenant :
\itemize{
\item \code{n_auto_adhpha_global} : identifiant de la pharmacie
\item \verb{CA xxxx} : le chiffre d'affaires pour l’année cible
}
}
\description{
Cette fonction retourne les pharmacies considérées comme \strong{complètes} sur 12 mois,
pour une année spécifique (par défaut l'année précédente), ainsi que leur chiffre d'affaires (CA).
}
\details{
\itemize{
\item Si \code{n = 0}, retourne les données pour l’année en cours - 1.
\item Si \code{n = 1}, retourne l’année en cours - 2, etc.
\item Jusqu’à \code{n = 2} (au-delà, un message est renvoyé).
}

Le nom de la colonne du CA est généré dynamiquement (ex : "CA 2023").
}
\examples{
\dontrun{
pharma_complete_12()       # Pour l’année N-1
pharma_complete_12(n = 1)  # Pour l’année N-2
}

}
